numeric: add float logic to numeric class not float

+++ lib/core_ext/numeric.rb
@@ -1,4 +1,4 @@
-class Float
+class Numeric
   def prime_number?
     (2...self).none?{ |n| (self / n) % 1 == 0 }
   end
